requirements.txt
streamlit
notion-client
python-dotenv
pandas
requests
google-generativeai
